Fix #3420: Separate MarkChaptersCompletedActivityTest into activity and fragment test files
